backend/requirements.txt
Django>=4.2
djangorestframework
channels
channels_redis
django-cors-headers
daphne
